move resource config files
